added 3d models

+++ clean.sh
@@ -7,14 +7,20 @@ do
         cd $i
         if [ -f "$FILE" ]; then # if file exists within current directory
             echo "$FILE in $i exist"
-            make clean 
-        else
-            cd ../
-            echo "adding Makefile to $i"
-            cp clean.txt $i
-            cd $i
-            mv clean.txt Makefile
             make clean
+        else
+            if [ $i == "./MicromouseV1_Software" ]; then # if file exists within current directory
+                echo "No need to add $FILE to $i"
+                echo "next..."
+            else
+                echo "$i"
+                cd ../
+                echo "adding Makefile to $i"
+                cp clean.txt $i
+                cd $i
+                mv clean.txt Makefile
+                make clean
+            fi
         fi
         cd ../
     fi
